Create exp-2.1-b.cpp
Create a class student having student uid and getnumber(),putnumber() as member
functions to get the values and display it.

Derive a class test having marks in different subjects and getmarks() and
putmarks() as member functions to get and display the values.

Derive another class sports from student class having sports score and
getscore(), putscore() as member functions to get and display the values.

Derive a class result from test and sports class and define a function display()
to calculate total marks.

Implement it with the object of result class. If it gives any error, resolve it
by adding the required functionality.