Overload iostream operators
Aim:- WAP to create a class Pairs. Objects of type Pairs can be used in any
situation where ordered pairs are needed. Our Task is to overload operator >>
and << so that objects of class Pairs are to be input and output in the form
(5,3) (5,-6) (-5,6) or (-5,-3).There is no need to implement any
constructor/method .